+++ src/main/java/se/lnu/application/processor/CollectionProcessor.java
@@ -53,20 +53,22 @@ public class CollectionProcessor implements Processor<CollectionDTO> {
     }
 
     @Override
-    public CollectionDTO update(CollectionDTO dto) {
+    public CollectionDTO update(CollectionDTO dto, AuthUser user) {
         // TODO add artifact removing logic
         CollectionEntity collectionEntity = collectionConverter.convertToEntity(dto);
+        checkPermission(collectionEntity.getUser().getId(), user.getId());
         collectionEntity = collectionDAO.update(collectionEntity);
         return collectionConverter.convertToDTO(collectionEntity);
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Long id, AuthUser user) {
         // TODO add deleting logic
         CollectionEntity collectionEntity = collectionDAO.get(id);
         if (collectionEntity == null) {
             throw new RecordNotFoundException(ErrorCode.COLLECTION_NOT_FOUND);
         }
+        checkPermission(collectionEntity.getUser().getId(), user.getId());
         collectionDAO.delete(collectionEntity);
     }
